+++ packages/@headlessui-react/src/index.ts
@@ -1,3 +1,5 @@
+import 'client-only'
+
 export * from './components/combobox/combobox'
 export * from './components/dialog/dialog'
 export * from './components/disclosure/disclosure'